+++ src/ts/markdown/setLute.ts
@@ -13,6 +13,7 @@ export const setLute = (options: ILuteOptions) => {
     lute.SetSetext(options.setext);
     lute.SetSanitize(options.sanitize);
     lute.SetChineseParagraphBeginningSpace(options.paragraphBeginningSpace);
+    lute.SetRenderListMarker(options.listMarker);
     if (options.lazyLoadImage) {
         lute.SetImageLazyLoading(options.lazyLoadImage);
     }