+++ src/ts/util/Options.ts
@@ -87,6 +87,7 @@ export class Options {
                 codeBlockPreview: true,
                 fixTermTypo: false,
                 footnotes: true,
+                listMarker: false,
                 sanitize: true,
                 setext: false,
                 theme: "light",